Adds canonification of required nodes
Splits canonify pass into two phases.  The first is where nodes are replaced by simpler nodes.  The second is where the DFAs are constructed.

In the first phase, we canonify required nodes by splitting them into reqmask/reqbit nodes.  The REQMASK node is AND’d with a PROPSET that matches each required property name.  The constraint on each match is a REQBIT node that sets a unique bit for each required property.

The tree is then resimplified, allowing PROPSETs to merge together.  Finally, in the second phase of the canonification, DFAs are built from the PROPSETs and the PROPSET/PROP_MATCH nodes are replaced with MATCH_SWITCH / MATCH_CASE nodes.