Adds small optimization which removes an op assembly edge case
in IR translation, we would always add a MATCH statement for any object.  If the object had no property matching constraints, the fsm attached would be NULL and only the default case would be relevant.  The op assembler would then have to recognize this case and generate code appropriately.

It’s much easier to change the emitted IR to not include a MATCH statement if there’s no need to match property names.  Instead, we add a CONSUME statement to consume the property name and then run the default case of the match.

This reduces the amount of IR generated and makes it easier for the op assembler to do its job.