Removes a source of non-determinism
libfsm sorts things by pointer value, so the results are non-deterministic.  This occasionally shows up when tests fail because MATCH_CASE nodes have been merged in a different order.

To compensate for this, we resort the children of AND/OR/XOR nodes on the second canonical pass.

NB: Currently the sort is only by type, and should be extended to correctly sort different nodes of the same type.