Fixes bug in cnode canonification
The way we copied the MATCH_CASE nodes was to iterate through the DFA end states, copy each node, and then replace the nodes on the MATCH_SWITCH cases list.

However, this runs into problems when several end states in the DFA could point to the same MATCH_CASE node.  This fixes that behavior in several ways:

1. After the DFA is initially built and minimized, we collect the MATCH_CASE nodes in a list.  We now keep track of nodes that have already been collected to ensure that the list has no duplicates.

2. Copying and subsequent modification is done on this unique list and then the opaque values of the states in the DFA are set to the new values.  This prevents double duplication (or worse) while still allowing several end states of the DFA to hold the same MATCH_CASE.

3. This also happens in the IR when we translate the MATCH_SWITCH/MATCH_CASE nodes into an IR.  Fixes this so the translation happens correctly.