Fixes items/additionalItems AST and conversion into cnodes
The items property can occur more than once and can overwrite previous settings.  This means that we can’t safely stash a non-array items constraint in the additional_items field of the ast.

To handle this, we add a keyword flag that indicates whether items has a non-array (singleton) value or not.  Translation of additionalItems into a cnode only occurs if items holds a singleton value.