Changes additional_properties to be a single schema
Instead of a set of schemas, additional properties is now a single schema.

I don’t see anywhere in the json schema spec that specifies what should happen if a validation keyword is repeated, but it doesn’t make a lot of sense to add additional schema constraints in this case to the original additionalProperties constraint.  Instead, we’ll just replace the additionalProperties constraint with a new one.