Adds all-case constraints to MATCH_SWITCH cnodes
Moves constraints on the string length (LENGTH_RANGE cnodes) so they’re children of the MATCH_SWITCH node.

Correctly processes this for strings in the IR.  Need to add processing for object properties, and potentially combine the cases.

This also sketches out how one would AND together the cases for several AND’ed MATCH_SWITCH nodes, although this is left unimplemented.