Initial change to MATCH_SWTICH string constraints
Work-in-progress commit.

To handle ORing of string conditions, we need to change how we handle string constraints in MATCH_SWITCH nodes.

This eliminates the contraints part of the match switch.  LENGTH_RANGE nodes are at the head of MATCH_CASE subnodes and the default case.

The cnode operations are now correct, although they need to be cleaned up.  **The IR translation is currently broken.**