Fixes issue merging mswitches
Previously, merging mswitches would combine endstates with different
constraints, a clear bug.  This was fixed by removing some of the
fsm_minimise calls, but this resulted in extra cases.  This commit
relies on a new two-dfa walking fsm_subtract that doesn’t generate a
complement and correctly combines the opaque values.