Adds “UNTOKEN” opcode, changes VM token handling to be a state machine
When handling array items, we have to peek at the next token to see if the array is ending or not.  Thus we need to fetch the token with a TOKEN instruction.  Previously, we just omitted the TOKEN instruction at the beginning of the FRAME for the array item, to avoid fetching two tokens.

However, with the id/ref stuff, the FRAME called might be defined elsewhere, and could call TOKEN.  Or the FRAME that omits the token might be ref’d elsewhere, and by not fetching a new token it would also have invalid results.

To handle this, we now call UNTOKEN before all array handling FRAMES.  This serves as an “ungetc” for the token.  The UNTOKEN opcode is encoded as  TOKEN instruction with -1 for the second argument.

This has also exposed the oddness of the ‘consumed’ and ‘nexttok’ members of struct jvst_vm.  I’ve replaced them with a single field that indicates the state of the jvst_vm token.  The ‘needtok’ member may not be redundant, but I need to think more about that.

NB: Also adds a debug option to the jvst cli to print tokens as they’re read