Fixes refs that refer to refs
Any cnode that’s the target of a reference is turned into a separate tree in the cnode forest.  During cnode -> IR translation, two things happen:

1) cnode REFs are converted directly into IR CALL_ID statements
2) Each cnode tree is translated separately.  We require that each tree begins with a FRAME.

One edge case was when the target of a cnode REF is another cnode REF.  In this case, the targetted REF was both translated into a CALL_ID statement and this CALL_ID statement was made a tree in the IR forest.  This violates our assumption that all IR trees are FRAMEs.

To fix this, we detect the situation and instead of making the targeted REF a root, we wrap it in a FRAME and make that frame the root.

All but one json schema test suite “ref” tests now pass.  The remaining one involves an external reference.