Adds a few special-cases to the VM to make UNIQUE work
So, some of this is pretty kludgy, but it’s a start.

The basic problem is that we need to feed tokens to the UNIQUE validator similar to how we SPLIT the validator for AND/OR cases.

However, the UNIQUE validator is a bit different in that it’s implemented as a pushdown automaton in C, and all we really need is to initialize the automaton and feed it tokens.

The complication here is how the automaton interacts with the current SPLIT architecture.  In the current SPLIT architecture, there’s no way to pass state between one invocation of the SPLIT and another.  We currently initialize the unique automaton in the parent VM, but there’s no clear way to pass it to a child VM.  The kludge here is to initialize the unique data on the child VM by detecting that the only instruction in the split is a UNIQUE_EVAL instruction.

This isn’t ideal, and there are several ways we could clean this up.  However, it seems to work (for now), and jvst has been blocked on this feature for a while, so I’m going to commit the kludge and hope to clean it up later (real soon now…).