Fixes failing custom/* tests, generates better code
The custom/* tests were failing for an odd reason.  In adding unique, I added a bunch of conditionals that often had NOPs in the branches.

When the IR was linearized, this would generate blocks that contained only a NOP instruction and a branch.

When the IR was flattened, the blocks would be rearranged and often the blocks with NOP/BRANCH could be simplified to a NOP.  However, there would still be a branch into these blocks.

During opcode assembly, a few things happen:
  1) “blocks” are converted into labels attached to particular instructions.
  2) NOP instructions are elided.
  3) Branches to blocks are emitted and the specific destination is fixed up later.

Keep in mind that branches always point to blocks in the flattened IR.  Let BLK_A be a block with only a NOP instruction.  The situation is basically this:

1) A branch tp BLK_A would be emitted with a note to fix up the destination later.

2) When BLK_A is emitted, the NOP instruction is skipped.  However, there are no remaining instructions, so nothing is emitted.

3) No instructions are emitted, so the label for BLK_A is not attached to an instruction.  Effectively, BLK_A never gets a destination address because it has no instructions.

4) When branch destinations are fixed up, BLK_A has no destination address and we have an error (an assertion violation in this case).

The solution turns out to be remarkably simple: don’t emit NOPs during linearization.  There’s no reason to, and jvst can emit better bytecode when they’re elided before block scheduling.  In particular, a bunch of empty blocks and useless jumps can be omitted.